% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Conductor}
\alias{Conductor}
\title{Create a "conductor" tour}
\description{
In addition to this page, you can also directly access the documentation
of shepherd.js here: \url{https://shepherdjs.dev/docs/index.html}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Conductor-new}{\code{Conductor$new()}}
\item \href{#method-Conductor-init}{\code{Conductor$init()}}
\item \href{#method-Conductor-start}{\code{Conductor$start()}}
\item \href{#method-Conductor-step}{\code{Conductor$step()}}
\item \href{#method-Conductor-clone}{\code{Conductor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conductor$new(exitOnEsc = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exitOnEsc}}{Allow closing the tour by pressing "Escape". Default is}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{Conductor} object.
}

\subsection{Returns}{
A \code{Conductor} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-init"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conductor$init(session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A valid Shiny session. If \code{NULL} (default), the function
attempts to get the session with \code{shiny::getDefaultReactiveDomain()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise \code{Conductor}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-start"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conductor$start(session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A valid Shiny session. If \code{NULL} (default), the function
attempts to get the session with \code{shiny::getDefaultReactiveDomain()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Start \code{Conductor}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-step"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-step}{}}}
\subsection{Method \code{step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conductor$step(title = NULL, text = NULL, el = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of the popover.}

\item{\code{text}}{Text of the popover.}

\item{\code{el}}{The element to highlight. It can be an id (for example \verb{#mynav}),
a class (for instance \code{.navbar}), or a general tag (for example \code{button}).
If \code{NULL} (default) or if the selector is not found, the popover will appear
in the center of the page.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add a step in a \code{Conductor} tour.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conductor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Conductor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conductor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
